Added additional features to the PrimeKG graph.

+++ neo4j_primekg/import_primekg.sh
@@ -1,9 +1,53 @@
 #!/bin/bash
 
+function download_data
+{
+    if [ -f "$PRIMEKG_CSV" ]; then
+        declare -g PRIMEKG_CSV_EXISTED=true
+    else
+        echo ">>> Downloading PrimeKG csv file..."
+        declare -g PRIMEKG_CSV_EXISTED=false
+        wget --no-clobber -O $PRIMEKG_CSV https://dataverse.harvard.edu/api/access/datafile/6180620
+    fi
+
+    if [ -f "$PRIMEKG_DRUG_NODE_FEATURES_TSV" ]; then
+        declare -g PRIMEKG_DRUG_FEATURES_EXISTED=true
+    else
+        echo ">>> Downloading PrimeKG drug features file..."
+        declare -g PRIMEKG_DRUG_FEATURES_EXISTED=false
+        wget --no-clobber -O $PRIMEKG_DRUG_NODE_FEATURES_TSV https://dataverse.harvard.edu/api/access/datafile/6180619
+    fi
+
+    if [ -f "$PRIMEKG_DISEASE_NODE_FEATURES_TSV" ]; then
+        declare -g PRIMEKG_DISEASE_FEATURES_EXISTED=true
+    else
+        echo ">>> Downloading PrimeKG disease features file..."
+        declare -g PRIMEKG_DISEASE_FEATURES_EXISTED=false
+        wget --no-clobber -O $PRIMEKG_DISEASE_NODE_FEATURES_TSV https://dataverse.harvard.edu/api/access/datafile/6180618
+    fi
+}
+
+function cleanup_data
+{
+    rm -r $PRIMEKG_CSVS_FOR_NEO4J
+
+    if [ "${PRIMEKG_CSV_EXISTED}" == "false" ]; then
+        echo ">>> Cleaning up PrimeKG csv file..."
+        rm $PRIMEKG_CSV
+    fi
+    if [ "${PRIMEKG_DRUG_FEATURES_EXISTED}" == "false" ]; then
+        echo ">>> Cleaning up PrimeKG csv file..."
+        rm $PRIMEKG_DRUG_NODE_FEATURES_TSV
+    fi
+    if [ "${PRIMEKG_DISEASE_FEATURES_EXISTED}" == "false" ]; then
+        echo ">>> Cleaning up PrimeKG csv file..."
+        rm $PRIMEKG_DISEASE_NODE_FEATURES_TSV
+    fi
+}
+
 function import_primekg
 {
-    echo ">>> Downloading PrimeKG csv file..."
-    wget --no-clobber -O $PRIMEKG_CSV https://dataverse.harvard.edu/api/access/datafile/6180620
+    download_data
 
     echo ">>> Processing PrimeKG csv file..."
     IMPORT_CMD=$(python3 $IMPORT_DIR/primekg_to_neo4j_csv.py)
@@ -22,11 +66,12 @@ function import_primekg
         exit 1
     fi
 
-    if [ "${DELETE_PRIMEKG_CSV}" == "true" ]; then
-        echo ">>> Cleaning up..."
-        rm -r /$IMPORT_DIR
-    fi
+    cleanup_data
 }
+
+if [ -e $DELETE_PRIMEKG_CSV ]; then
+    echo "WARNING: DELETE_PRIMEKG_CSV is deprecated. This now gets handled automatically."
+fi
 
 SETUP_DONE_MARKER="/data/prime_kg_is_imported_to_neo4j"
 if [ ! -e $SETUP_DONE_MARKER ]; then
